+++ service/globalaccelerator/go_module_metadata.go
@@ -3,4 +3,4 @@
 package globalaccelerator
 
 // goModuleVersion is the tagged release for this module
-const goModuleVersion = "1.35.9"
+const goModuleVersion = "1.35.10"