+++ service/migrationhubrefactorspaces/go_module_metadata.go
@@ -3,4 +3,4 @@
 package migrationhubrefactorspaces
 
 // goModuleVersion is the tagged release for this module
-const goModuleVersion = "1.25.16"
+const goModuleVersion = "1.25.17"