+++ service/transcribe/api_op_GetMedicalVocabulary.go
@@ -8,7 +8,9 @@ import (
 	awsmiddleware "github.com/aws/aws-sdk-go-v2/aws/middleware"
 	"github.com/aws/aws-sdk-go-v2/service/transcribe/types"
 	"github.com/aws/smithy-go/middleware"
+	smithytime "github.com/aws/smithy-go/time"
 	smithyhttp "github.com/aws/smithy-go/transport/http"
+	smithywaiter "github.com/aws/smithy-go/waiter"
 	"time"
 )
 
@@ -180,6 +182,201 @@ func (c *Client) addOperationGetMedicalVocabularyMiddlewares(stack *middleware.S
 	}
 	return nil
 }
+
+// MedicalVocabularyReadyWaiterOptions are waiter options for
+// MedicalVocabularyReadyWaiter
+type MedicalVocabularyReadyWaiterOptions struct {
+
+	// Set of options to modify how an operation is invoked. These apply to all
+	// operations invoked for this client. Use functional options on operation call to
+	// modify this list for per operation behavior.
+	//
+	// Passing options here is functionally equivalent to passing values to this
+	// config's ClientOptions field that extend the inner client's APIOptions directly.
+	APIOptions []func(*middleware.Stack) error
+
+	// Functional options to be passed to all operations invoked by this client.
+	//
+	// Function values that modify the inner APIOptions are applied after the waiter
+	// config's own APIOptions modifiers.
+	ClientOptions []func(*Options)
+
+	// MinDelay is the minimum amount of time to delay between retries. If unset,
+	// MedicalVocabularyReadyWaiter will use default minimum delay of 10 seconds. Note
+	// that MinDelay must resolve to a value lesser than or equal to the MaxDelay.
+	MinDelay time.Duration
+
+	// MaxDelay is the maximum amount of time to delay between retries. If unset or
+	// set to zero, MedicalVocabularyReadyWaiter will use default max delay of 120
+	// seconds. Note that MaxDelay must resolve to value greater than or equal to the
+	// MinDelay.
+	MaxDelay time.Duration
+
+	// LogWaitAttempts is used to enable logging for waiter retry attempts
+	LogWaitAttempts bool
+
+	// Retryable is function that can be used to override the service defined
+	// waiter-behavior based on operation output, or returned error. This function is
+	// used by the waiter to decide if a state is retryable or a terminal state.
+	//
+	// By default service-modeled logic will populate this option. This option can
+	// thus be used to define a custom waiter state with fall-back to service-modeled
+	// waiter state mutators.The function returns an error in case of a failure state.
+	// In case of retry state, this function returns a bool value of true and nil
+	// error, while in case of success it returns a bool value of false and nil error.
+	Retryable func(context.Context, *GetMedicalVocabularyInput, *GetMedicalVocabularyOutput, error) (bool, error)
+}
+
+// MedicalVocabularyReadyWaiter defines the waiters for MedicalVocabularyReady
+type MedicalVocabularyReadyWaiter struct {
+	client GetMedicalVocabularyAPIClient
+
+	options MedicalVocabularyReadyWaiterOptions
+}
+
+// NewMedicalVocabularyReadyWaiter constructs a MedicalVocabularyReadyWaiter.
+func NewMedicalVocabularyReadyWaiter(client GetMedicalVocabularyAPIClient, optFns ...func(*MedicalVocabularyReadyWaiterOptions)) *MedicalVocabularyReadyWaiter {
+	options := MedicalVocabularyReadyWaiterOptions{}
+	options.MinDelay = 10 * time.Second
+	options.MaxDelay = 120 * time.Second
+	options.Retryable = medicalVocabularyReadyStateRetryable
+
+	for _, fn := range optFns {
+		fn(&options)
+	}
+	return &MedicalVocabularyReadyWaiter{
+		client:  client,
+		options: options,
+	}
+}
+
+// Wait calls the waiter function for MedicalVocabularyReady waiter. The
+// maxWaitDur is the maximum wait duration the waiter will wait. The maxWaitDur is
+// required and must be greater than zero.
+func (w *MedicalVocabularyReadyWaiter) Wait(ctx context.Context, params *GetMedicalVocabularyInput, maxWaitDur time.Duration, optFns ...func(*MedicalVocabularyReadyWaiterOptions)) error {
+	_, err := w.WaitForOutput(ctx, params, maxWaitDur, optFns...)
+	return err
+}
+
+// WaitForOutput calls the waiter function for MedicalVocabularyReady waiter and
+// returns the output of the successful operation. The maxWaitDur is the maximum
+// wait duration the waiter will wait. The maxWaitDur is required and must be
+// greater than zero.
+func (w *MedicalVocabularyReadyWaiter) WaitForOutput(ctx context.Context, params *GetMedicalVocabularyInput, maxWaitDur time.Duration, optFns ...func(*MedicalVocabularyReadyWaiterOptions)) (*GetMedicalVocabularyOutput, error) {
+	if maxWaitDur <= 0 {
+		return nil, fmt.Errorf("maximum wait time for waiter must be greater than zero")
+	}
+
+	options := w.options
+	for _, fn := range optFns {
+		fn(&options)
+	}
+
+	if options.MaxDelay <= 0 {
+		options.MaxDelay = 120 * time.Second
+	}
+
+	if options.MinDelay > options.MaxDelay {
+		return nil, fmt.Errorf("minimum waiter delay %v must be lesser than or equal to maximum waiter delay of %v.", options.MinDelay, options.MaxDelay)
+	}
+
+	ctx, cancelFn := context.WithTimeout(ctx, maxWaitDur)
+	defer cancelFn()
+
+	logger := smithywaiter.Logger{}
+	remainingTime := maxWaitDur
+
+	var attempt int64
+	for {
+
+		attempt++
+		apiOptions := options.APIOptions
+		start := time.Now()
+
+		if options.LogWaitAttempts {
+			logger.Attempt = attempt
+			apiOptions = append([]func(*middleware.Stack) error{}, options.APIOptions...)
+			apiOptions = append(apiOptions, logger.AddLogger)
+		}
+
+		out, err := w.client.GetMedicalVocabulary(ctx, params, func(o *Options) {
+			baseOpts := []func(*Options){
+				addIsWaiterUserAgent,
+			}
+			o.APIOptions = append(o.APIOptions, apiOptions...)
+			for _, opt := range baseOpts {
+				opt(o)
+			}
+			for _, opt := range options.ClientOptions {
+				opt(o)
+			}
+		})
+
+		retryable, err := options.Retryable(ctx, params, out, err)
+		if err != nil {
+			return nil, err
+		}
+		if !retryable {
+			return out, nil
+		}
+
+		remainingTime -= time.Since(start)
+		if remainingTime < options.MinDelay || remainingTime <= 0 {
+			break
+		}
+
+		// compute exponential backoff between waiter retries
+		delay, err := smithywaiter.ComputeDelay(
+			attempt, options.MinDelay, options.MaxDelay, remainingTime,
+		)
+		if err != nil {
+			return nil, fmt.Errorf("error computing waiter delay, %w", err)
+		}
+
+		remainingTime -= delay
+		// sleep for the delay amount before invoking a request
+		if err := smithytime.SleepWithContext(ctx, delay); err != nil {
+			return nil, fmt.Errorf("request cancelled while waiting, %w", err)
+		}
+	}
+	return nil, fmt.Errorf("exceeded max wait time for MedicalVocabularyReady waiter")
+}
+
+func medicalVocabularyReadyStateRetryable(ctx context.Context, input *GetMedicalVocabularyInput, output *GetMedicalVocabularyOutput, err error) (bool, error) {
+
+	if err == nil {
+		v1 := output.VocabularyState
+		expectedValue := "READY"
+		var pathValue string
+		pathValue = string(v1)
+		if pathValue == expectedValue {
+			return false, nil
+		}
+	}
+
+	if err == nil {
+		v1 := output.VocabularyState
+		expectedValue := "FAILED"
+		var pathValue string
+		pathValue = string(v1)
+		if pathValue == expectedValue {
+			return false, fmt.Errorf("waiter state transitioned to Failure")
+		}
+	}
+
+	if err != nil {
+		return false, err
+	}
+	return true, nil
+}
+
+// GetMedicalVocabularyAPIClient is a client that implements the
+// GetMedicalVocabulary operation.
+type GetMedicalVocabularyAPIClient interface {
+	GetMedicalVocabulary(context.Context, *GetMedicalVocabularyInput, ...func(*Options)) (*GetMedicalVocabularyOutput, error)
+}
+
+var _ GetMedicalVocabularyAPIClient = (*Client)(nil)
 
 func newServiceMetadataMiddleware_opGetMedicalVocabulary(region string) *awsmiddleware.RegisterServiceMetadata {
 	return &awsmiddleware.RegisterServiceMetadata{