+++ service/workspaces/deserializers.go
@@ -20933,6 +20933,19 @@ func awsAwsjson11_deserializeOpDocumentDescribeCustomWorkspaceImageImportOutput(
 				}
 			}
 
+		case "ProgressPercentage":
+			if value != nil {
+				jtv, ok := value.(json.Number)
+				if !ok {
+					return fmt.Errorf("expected Percentage to be json.Number, got %T instead", value)
+				}
+				i64, err := jtv.Int64()
+				if err != nil {
+					return err
+				}
+				sv.ProgressPercentage = ptr.Int32(int32(i64))
+			}
+
 		case "State":
 			if value != nil {
 				jtv, ok := value.(string)
@@ -20940,6 +20953,15 @@ func awsAwsjson11_deserializeOpDocumentDescribeCustomWorkspaceImageImportOutput(
 					return fmt.Errorf("expected CustomWorkspaceImageImportState to be of type string, got %T instead", value)
 				}
 				sv.State = types.CustomWorkspaceImageImportState(jtv)
+			}
+
+		case "StateMessage":
+			if value != nil {
+				jtv, ok := value.(string)
+				if !ok {
+					return fmt.Errorf("expected WorkflowStateMessage to be of type string, got %T instead", value)
+				}
+				sv.StateMessage = ptr.String(jtv)
 			}
 
 		default: