+++ service/fis/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.37.16 (2026-01-09)
+
+* **Dependency Update**: Updated to the latest SDK module versions
+
 # v1.37.15 (2025-12-08)
 
 * **Dependency Update**: Updated to the latest SDK module versions