+++ service/sagemakermetrics/CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.18.19 (2026-01-09)
+
+* **Dependency Update**: Updated to the latest SDK module versions
+
 # v1.18.18 (2025-12-09)
 
 * No change notes available for this release.